+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#3306](https://github.com/bbatsov/rubocop/issues/3306): Add autocorrection for `Style/EachWithObject`. ([@owst][])
 * Add new `Style/TernaryParentheses` cop. ([@drenmi][])
 * [#3136](https://github.com/bbatsov/rubocop/issues/3136): Add config for `UselessAccessModifier` so it can be made aware of ActiveSupport's `concerning` and `class_methods` methods. ([@maxjacobson][])
+* [#3128](https://github.com/bbatsov/rubocop/issues/3128): Add new `Rails/SaveBang` cop. ([@QuinnHarris][])
 
 ### Bug fixes
 
@@ -2288,3 +2289,4 @@
 [@Tei]: https://github.com/Tei
 [@haziqhafizuddin]: https://github.com/haziqhafizuddin
 [@dvandersluis]: https://github.com/dvandersluis
+[@QuinnHarris]: https://github.com/QuinnHarris